+++ packages/react/src/components/link/index.ts
@@ -5,15 +5,8 @@ export * from "./link.style"
 export {
   LinkBox,
   LinkBoxPropsContext,
+  LinkOverlay,
   useLinkBoxPropsContext,
 } from "./link-box"
-export type { LinkBoxProps } from "./link-box"
+export type { LinkBoxProps, LinkOverlayProps } from "./link-box"
 export * from "./link-box.style"
-
-export {
-  LinkOverlay,
-  LinkOverlayPropsContext,
-  useLinkOverlayPropsContext,
-} from "./link-overlay"
-export type { LinkOverlayProps } from "./link-overlay"
-export * from "./link-overlay.style"
