+++ packages/react/src/components/link/link-box.style.ts
@@ -2,6 +2,7 @@ import { defineComponentSlotStyle } from "../../core"
 
 export const linkBoxStyle = defineComponentSlotStyle({
   base: {
+    container: {},
     overlay: {
       outline: 0,
       _focusVisible: {
@@ -10,7 +11,6 @@ export const linkBoxStyle = defineComponentSlotStyle({
         },
       },
     },
-    root: {},
   },
 })
